refactor(frontend): extract Buttons component into StepperButtons and update imports
