+++ apps/competition-manager-web/src/Components/index.tsx
@@ -17,3 +17,4 @@ export * from './Icons';
 export * from './CircleButton';
 export * from './AccountCircle';
 export * from './ErrorFallback';
+export * from './StepperButtons';
